Address review comments, rename noddy_rebuild to simple_rebuild
