+++ mayastor/src/bdev/nexus/mod.rs
@@ -20,7 +20,7 @@ pub mod nexus_module;
 pub mod nexus_nbd;
 pub mod nexus_rpc;
 pub mod nexus_share;
-pub mod nexus_noddy_rebuild;
+pub mod nexus_simple_rebuild;
 
 /// public function which simply calls register module
 pub fn register_module() {
